feat(android): display filtered app in expanded view
