+++ composeApp/src/androidMain/kotlin/fr/husi/ui/AppListViewModel.kt
@@ -12,7 +12,8 @@ import kotlinx.coroutines.launch
 @Immutable
 internal data class AppListUiState(
     val isLoading: Boolean = false,
-    val apps: List<ProxiedApp> = emptyList(), // sorted, for show
+    val apps: List<ProxiedApp> = emptyList(), // sorted, full
+    val filteredApps: List<ProxiedApp> = emptyList(), // sorted, filtered by search
     val snackbarMessage: StringOrRes? = null,
 )
 
@@ -40,8 +41,8 @@ internal class AppListViewModel : BaseAppListViewModel() {
         }
     }
 
-    override fun updateApps(apps: List<ProxiedApp>, isLoading: Boolean) {
-        _uiState.update { it.copy(isLoading = isLoading, apps = apps) }
+    override fun updateApps(apps: List<ProxiedApp>, filteredApps: List<ProxiedApp>, isLoading: Boolean) {
+        _uiState.update { it.copy(isLoading = isLoading, apps = apps, filteredApps = filteredApps) }
     }
 
     override fun updateSnackbar(message: StringOrRes?) {
